Add user and thought models, routes and auth middleware
